Bayesian Network
Developed various Bayesian Network using crimes with 911 without demographic data, since it only takes discrete values (categorical). The difference among the models are how the conditional probabilities are calculated/approximated. For our dataset, the only feasible ones are "exact-dp" and "chow-liu". Got an accuracy of 0.67 on a small test set.